requirements.txt
aiogram==3.*
Flask==2.3.3
Jinja2==3.1.3
python-dotenv
psycopg2-binary
